requirements.txt
mitmproxy
PyYaml
sanic
flask
Sanic-Dispatcher